backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.10
pyjwt>=2.9.0
python-dotenv>=1.0.1
pydantic-settings>=2.6.0
bcrypt>=4.0.0
email-validator>=2.0.0
openai>=1.50.0
openai-agents>=0.1.0
mcp>=1.0.0
